requirements.txt
streamlit
langchain
langchain-openai
openai
pypdf2
faiss-cpu
python-dotenv
tiktoken